+++ update/update_phpbb_to_2022.php
@@ -85,7 +85,6 @@ hr	{ height: 0px; border: solid #D1D7DC 0px; border-top-width: 1px;}
 .ok {color:green}
 
 /* Import the fancy styles for IE only (NS4.x doesn't use the @import function) */
-@import url("../templates/subSilver/formIE.css");
 -->
 </style>
 </head>
@@ -95,7 +94,7 @@ hr	{ height: 0px; border: solid #D1D7DC 0px; border-top-width: 1px;}
 	<tr>
 		<td><table width="100%" border="0" cellspacing="0" cellpadding="0">
 			<tr>
-				<td><img src="../templates/subSilver/images/logo_phpBB.gif" border="0" alt="Forum Home" vspace="1" /></td>
+				<td><img src="../templates/fisubsilversh/images/logo_phpBB_med.gif" border="0" alt="Forum Home" vspace="1" /></td>
 <td align="center" width="100%" valign="middle"><span class="maintitle">Updating phpBB to 2.0.22</span></td>
 			</tr>
 		</table></td>
@@ -736,7 +735,7 @@ switch ($row['config_value'])
 	case '':
 		$sql = "SELECT themes_id
 			FROM " . THEMES_TABLE . "
-			WHERE template_name = 'subSilver'";
+			WHERE template_name = 'fisubsilversh'";
 		$result = _sql($sql, $errored, $error_ary);
 
 		if ($row = $db->sql_fetchrow($result))
@@ -744,7 +743,7 @@ switch ($row['config_value'])
 			$theme_id = $row['themes_id'];
 
 			$sql = "UPDATE " . THEMES_TABLE . "
-				SET head_stylesheet = 'subSilver.css', body_background = '', body_bgcolor = 'E5E5E5', body_text = '000000', body_link = '006699', body_vlink = '5493B4', body_alink = '', body_hlink = 'DD6900', tr_color1 = 'EFEFEF', tr_color2 = 'DEE3E7', tr_color3 = 'D1D7DC', tr_class1 = '', tr_class2 = '', tr_class3 = '', th_color1 = '98AAB1', th_color2 = '006699', th_color3 = 'FFFFFF', th_class1 = 'cellpic1.gif', th_class2 = 'cellpic3.gif', th_class3 = 'cellpic2.jpg', td_color1 = 'FAFAFA', td_color2 = 'FFFFFF', td_color3 = '', td_class1 = 'row1', td_class2 = 'row2', td_class3 = '', fontface1 = 'Verdana, Arial, Helvetica, sans-serif', fontface2 = 'Trebuchet MS', fontface3 = 'Courier, ''Courier New'', sans-serif', fontsize1 = 10, fontsize2 = 11, fontsize3 = 12, fontcolor1 = '444444', fontcolor2 = '006600', fontcolor3 = 'FFA34F', span_class1 = '', span_class2 = '', span_class3 = ''
+				SET head_stylesheet = 'fisubsilversh.css'
 				WHERE themes_id = $theme_id";
 			_sql($sql, $errored, $error_ary);